requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
email-validator==2.1.0
Pillow==11.2.1
pytest==7.4.2
